src/main/java/lv/javaguru/student_andrei_de/Homework/lesson5/ArrayUtils.java
package main.java.lv.javaguru.student_andrei_de.Homework.lesson5;

import java.util.Random;

class ArrayUtils {

    public int[] createArray(int arrayLength) {
        return new int[arrayLength];

    }

    public void fillArrayWithRandomNumbers(int[] array) {
        Random random = new Random();
        for (int i = 0; i < array.length; i++) {
            array[i] = random.nextInt(100);
        }
    }


    public int findMaxNumber(int[] array) {
        int max = array[0];
        for (int i = 0; i < array.length; i++) {
            if (array[i] > max) {
            max = array[i];
        }
    }
    return max;

       }

    public int findMinNumber(int[] array) {
        int min = array[0];
        for (int i = 0; i < array.length; i++) {
            if (array[i] < min) {
                min = array[i];
            }
        }
        return min;
    }
}

